Add references to local repository
Apparently, the Commons parent POM is located in the local
repository, and children projects would erroneously look
for this POM in the public repositories during builds.

Make sure to add references to the local repository to
correctly point the children projects there to look for
the parent POM (and other related dependencies).