Skip installation during Travis build
It was found out that goals that were skipped
(-D<plugin goal>.skip=true) would not fetch dependencies,
thus leaving the cache with incomplete dependencies having
to be downloaded either during the testing or the deployment
phases.

To solve this problem, it was decided to try and lump all
dependency fetching into the testing phase, while still
skipping over plugins not needed for deployment.